+++ src/Cache/CacheKeyDirective.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Nuwave\Lighthouse\Schema\Directives;
+namespace Nuwave\Lighthouse\Cache;
 
+use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Support\Contracts\Directive;
 
 class CacheKeyDirective extends BaseDirective implements Directive