+++ tests/Integration/Schema/Directives/ThrottleDirectiveTest.php
@@ -88,7 +88,9 @@ final class ThrottleDirectiveTest extends TestCase
         {
             foo
         }
-        ')->assertGraphQLErrorMessage(RateLimitException::MESSAGE);
+        ')->assertGraphQLError(
+            new RateLimitException('Query.foo')
+        );
     }
 
     public function testInlineLimiter(): void
@@ -113,6 +115,8 @@ final class ThrottleDirectiveTest extends TestCase
         {
             foo
         }
-        ')->assertGraphQLErrorMessage(RateLimitException::MESSAGE);
+        ')->assertGraphQLError(
+            new RateLimitException('Query.foo')
+        );
     }
 }